post-service/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-filter==23.5
celery==5.3.4
redis==5.0.1
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0 